requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
spacy==3.7.2
transformers==4.35.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pymongo==4.6.0
python-docx==1.0.1
PyPDF2==3.0.1
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pydantic==2.5.2
email-validator==2.1.0.post1
twilio==8.10.0
boto3==1.29.3 